fib_trie: Push assignment of child to parent down into inflate/halve
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Alexander Duyck <alexander.h.duyck@redhat.com>
commit ff181ed8768fa57fbda40fa97914f27fe275129b

This change makes it so that the assignment of the tnode to the parent is
handled directly within whatever function is currently handling the node be
it inflate, halve, or resize.  By doing this we can avoid some of the need
to set NULL pointers in the tree while we are resizing the subnodes.

	Signed-off-by: Alexander Duyck <alexander.h.duyck@redhat.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit ff181ed8768fa57fbda40fa97914f27fe275129b)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>