pppoe: Use workqueue to die properly when a PADT is received
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Simon Farnsworth <simon@farnz.org.uk>
commit 287f3a943fef58c5c73e42545169443be379222f

When a PADT frame is received, the socket may not be in a good state to
close down the PPP interface. The current implementation handles this by
simply blocking all further PPP traffic, and hoping that the lack of traffic
will trigger the user to investigate.

Use schedule_work to get to a process context from which we clear down the
PPP interface, in a fashion analogous to hangup on a TTY-based PPP
interface. This causes pppd to disconnect immediately, and allows tools to
take immediate corrective action.

Note that pppd's rp_pppoe.so plugin has code in it to disable the session
when it disconnects; however, as a consequence of this patch, the session is
already disabled before rp_pppoe.so is asked to disable the session. The
result is a harmless error message:

Failed to disconnect PPPoE socket: 114 Operation already in progress

This message is safe to ignore, as long as the error is 114 Operation
already in progress; in that specific case, it means that the PPPoE session
has already been disabled before pppd tried to disable it.

	Signed-off-by: Simon Farnsworth <simon@farnz.org.uk>
	Tested-by: Dan Williams <dcbw@redhat.com>
	Tested-by: Christoph Schulz <develop@kristov.de>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 287f3a943fef58c5c73e42545169443be379222f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>