bdi: reimplement bdev_inode_switch_bdi()
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Tejun Heo <tj@kernel.org>
commit 018a17bdc8658ad448497c84d4ba21b6985820ec

A block_device may be attached to different gendisks and thus
different bdis over time.  bdev_inode_switch_bdi() is used to switch
the associated bdi.  The function assumes that the inode could be
dirty and transfers it between bdis if so.  This is a bit nasty in
that it reaches into bdi internals.

This patch reimplements the function so that it writes out the inode
if dirty.  This is a lot simpler and can be implemented without
exposing bdi internals.

	Signed-off-by: Tejun Heo <tj@kernel.org>
	Cc: Alexander Viro <viro@zeniv.linux.org.uk>
	Signed-off-by: Jens Axboe <axboe@fb.com>
(cherry picked from commit 018a17bdc8658ad448497c84d4ba21b6985820ec)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>