tools lib traceevent: Copy trace_clock and free it
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Steven Rostedt (Red Hat) <rostedt@goodmis.org>
commit 99ad1417db2e83530b88e3eff1a40285d6ebfee9

The pevent->trace_clock should not be a direct pointer to what was
given. It should be copied and freed.

Note, valgrind pointed this out when a caller passed in a pointer that
needed to be freed and it never was. Ideally, pevent should copy it
(which this change does), and free the copy. It's up to the caller to
free the clock string passed in.

	Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
	Cc: Andrew Morton <akpm@linux-foundation.org>
	Cc: Jiri Olsa <jolsa@redhat.com>
	Cc: Namhyung Kim <namhyung@kernel.org>
Link: http://lkml.kernel.org/r/20150324135922.695906738@goodmis.org
	Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>
(cherry picked from commit 99ad1417db2e83530b88e3eff1a40285d6ebfee9)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>