tools lib traceevent: Honor operator priority
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Namhyung Kim <namhyung@kernel.org>
commit 3201f0dc42f7fad9387afc4692cea3d0c730cba2

Currently it ignores operator priority and just sets processed args as a
right operand.  But it could result in priority inversion in case that
the right operand is also a operator arg and its priority is lower.

For example, following print format is from new kmem events.

  "page=%p", REC->pfn != -1UL ? (((struct page *)(0xffffea0000000000UL)) + (REC->pfn)) : ((void *)0)

But this was treated as below:

  REC->pfn != ((null - 1UL) ? ((struct page *)0xffffea0000000000UL + REC->pfn) : (void *) 0)

In this case, the right arg was '?' operator which has lower priority.
But it just sets the whole arg so making the output confusing - page was
always 0 or 1 since that's the result of logical operation.

With this patch, it can handle it properly like following:

  ((REC->pfn != (null - 1UL)) ? ((struct page *)0xffffea0000000000UL + REC->pfn) : (void *) 0)

	Signed-off-by: Namhyung Kim <namhyung@kernel.org>
	Acked-by: Steven Rostedt <rostedt@goodmis.org>
	Cc: David Ahern <dsahern@gmail.com>
	Cc: Jiri Olsa <jolsa@redhat.com>
	Cc: Joonsoo Kim <js1304@gmail.com>
	Cc: Minchan Kim <minchan@kernel.org>
	Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
	Cc: linux-mm@kvack.org
Link: http://lkml.kernel.org/r/1428298576-9785-10-git-send-email-namhyung@kernel.org
[ Replaced 'swap' with 'rotate' in a comment as requested by Steve and agreed by Namhyung ]
	Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>
(cherry picked from commit 3201f0dc42f7fad9387afc4692cea3d0c730cba2)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>