ext4: ignore journal checksum on remount; don't fail
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Eric Sandeen <sandeen@redhat.com>
commit 2d5b86e048780c5efa7f7d9708815555919e7b05

As of v3.18, ext4 started rejecting a remount which changes the
journal_checksum option.

Prior to that, it was simply ignored; the problem here is that
if someone has this in their fstab for the root fs, now the box
fails to boot properly, because remount of root with the new options
will fail, and the box proceeds with a readonly root.

I think it is a little nicer behavior to accept the option, but
warn that it's being ignored, rather than failing the mount,
but that might be a subjective matter...

	Reported-by: Cónräd <conradsand.arma@gmail.com>
	Signed-off-by: Eric Sandeen <sandeen@redhat.com>
	Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit 2d5b86e048780c5efa7f7d9708815555919e7b05)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>