KVM: x86: Wrong flags on CMPS and SCAS emulation
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [kvm] x86: Wrong flags on CMPS and SCAS emulation (Bandan Das) [1209995]
Rebuild_FUZZ: 94.51%
commit-author Nadav Amit <namit@cs.technion.ac.il>
commit 5aca3722362689a679419dc9b3e426921c70ef16

CMPS and SCAS instructions are evaluated in the wrong order.  For reference (of
CMPS), see http://www.fermimn.gov.it/linux/quarta/x86/cmps.htm : "Note that the
direction of subtraction for CMPS is [SI] - [DI] or [ESI] - [EDI]. The left
operand (SI or ESI) is the source and the right operand (DI or EDI) is the
destination. This is the reverse of the usual Intel convention in which the
left operand is the destination and the right operand is the source."

Introducing em_cmp_r for this matter that performs comparison in reverse order
using fastop infrastructure to avoid a wrapper function.

	Signed-off-by: Nadav Amit <namit@cs.technion.ac.il>
	Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 5aca3722362689a679419dc9b3e426921c70ef16)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>