NVMe: Fix filesystem sync deadlock on removal
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Keith Busch <keith.busch@intel.com>
commit 302c6727e5eb4d0be0e02d077b65feb3e73ea254

This changes the order of deleting the gendisks so it happens after the
nvme IO queues are freed. If a device is removed while a filesystem has
associated dirty data, the removal will wait on these to complete before
proceeding from del_gendisk, which could have caused deadlock before.

The implication of this is that an orderly removal of a responsive
device won't necessarily wait for dirty data to be written, but we are
not guaranteed the device is even going to respond at this point either.

	Signed-off-by: Keith Busch <keith.busch@intel.com>
	Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
	Signed-off-by: Jens Axboe <axboe@fb.com>
(cherry picked from commit 302c6727e5eb4d0be0e02d077b65feb3e73ea254)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>