drivers/base: provide an infrastructure for componentised subsystems
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [base] provide an infrastructure for componentised subsystems (Rob Clark) [1195471]
Rebuild_FUZZ: 88.52%
commit-author Russell King <rmk+kernel@arm.linux.org.uk>
commit 2a41e6070dd7ef539d0f3b1652b4839d04378e11

Subsystems such as ALSA, DRM and others require a single card-level
device structure to represent a subsystem.  However, firmware tends to
describe the individual devices and the connections between them.

Therefore, we need a way to gather up the individual component devices
together, and indicate when we have all the component devices.

We do this in DT by providing a "superdevice" node which specifies
the components, eg:

	imx-drm {
		compatible = "fsl,drm";
		crtcs = <&ipu1>;
		connectors = <&hdmi>;
	};

The superdevice is declared into the component support, along with the
subcomponents.  The superdevice receives callbacks to locate the
subcomponents, and identify when all components are present.  At this
point, we bind the superdevice, which causes the appropriate subsystem
to be initialised in the conventional way.

When any of the components or superdevice are removed from the system,
we unbind the superdevice, thereby taking the subsystem down.

	Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 2a41e6070dd7ef539d0f3b1652b4839d04378e11)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>