KVM: PPC: Book3S HV: Fix computation of tlbie operand
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [kvm] ppc: book3s hv: Fix computation of tlbie operand (Thomas Huth) [1226884 1227323]
Rebuild_FUZZ: 95.05%
commit-author Paul Mackerras <paulus@samba.org>
commit d506735b1a3c78e2efb9dc4019c76e9d3938a160

The B (segment size) field in the RB operand for the tlbie
instruction is two bits, which we get from the top two bits of
the first doubleword of the HPT entry to be invalidated.  These
bits go in bits 8 and 9 of the RB operand (bits 54 and 55 in IBM
bit numbering).

The compute_tlbie_rb() function gets these bits as v >> (62 - 8),
which is not correct as it will bring in the top 10 bits, not
just the top two.  These extra bits could corrupt the AP, AVAL
and L fields in the RB value.  To fix this we shift right 62 bits
and then shift left 8 bits, so we only get the two bits of the
B field.

The first doubleword of the HPT entry is under the control of the
guest kernel.  In fact, Linux guests will always put zeroes in bits
54 -- 61 (IBM bits 2 -- 9), but we should not rely on guests doing
this.

	Signed-off-by: Paul Mackerras <paulus@samba.org>
	Reviewed-by: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>
	Signed-off-by: Alexander Graf <agraf@suse.de>
(cherry picked from commit d506735b1a3c78e2efb9dc4019c76e9d3938a160)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>