FS-Cache: Add interface to check consistency of a cached object
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author David Howells <dhowells@redhat.com>
commit da9803bc8812f5bd3b26baaa90e515b843c65ff7

Extend the fscache netfs API so that the netfs can ask as to whether a cache
object is up to date with respect to its corresponding netfs object:

	int fscache_check_consistency(struct fscache_cookie *cookie)

This will call back to the netfs to check whether the auxiliary data associated
with a cookie is correct.  It returns 0 if it is and -ESTALE if it isn't; it
may also return -ENOMEM and -ERESTARTSYS.

The backends now have to implement a mandatory operation pointer:

	int (*check_consistency)(struct fscache_object *object)

that corresponds to the above API call.  FS-Cache takes care of pinning the
object and the cookie in memory and managing this call with respect to the
object state.

Original-author: Hongyi Jia <jiayisuse@gmail.com>
	Signed-off-by: David Howells <dhowells@redhat.com>
cc: Hongyi Jia <jiayisuse@gmail.com>
cc: Milosz Tanski <milosz@adfin.com>
(cherry picked from commit da9803bc8812f5bd3b26baaa90e515b843c65ff7)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>