vfio/pci: Restore MSIx message prior to enabling
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Gavin Shan <gwshan@linux.vnet.ibm.com>
commit b8f02af096b1fc9fd46680cbe55214e477eb76d3

The MSIx vector table lives in device memory, which may be cleared as
part of a backdoor device reset. This is the case on the IBM IPR HBA
when the BIST is run on the device. When assigned to a QEMU guest,
the guest driver does a pci_save_state(), issues a BIST, then does a
pci_restore_state(). The BIST clears the MSIx vector table, but due
to the way interrupts are configured the pci_restore_state() does not
restore the vector table as expected. Eventually this results in an
EEH error on Power platforms when the device attempts to signal an
interrupt with the zero'd table entry.

Fix the problem by restoring the host cached MSI message prior to
enabling each vector.

	Reported-by: Wen Xiong <wenxiong@linux.vnet.ibm.com>
	Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
	Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
(cherry picked from commit b8f02af096b1fc9fd46680cbe55214e477eb76d3)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>