IB/core: disallow registering 0-sized memory region
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Yann Droneaud <ydroneaud@opteya.com>
commit 8abaae62f3fdead8f4ce0ab46b4ab93dee39bab2

If ib_umem_get() is called with a size equal to 0 and an
non-page aligned address, one page will be pinned and a
0-sized umem will be returned to the caller.

This should not be allowed: it's not expected for a memory
region to have a size equal to 0.

This patch adds a check to explicitly refuse to register
a 0-sized region.

Link: http://mid.gmane.org/cover.1428929103.git.ydroneaud@opteya.com
	Cc: <stable@vger.kernel.org>
	Cc: Shachar Raindel <raindel@mellanox.com>
	Cc: Jack Morgenstein <jackm@mellanox.com>
	Cc: Or Gerlitz <ogerlitz@mellanox.com>
	Signed-off-by: Yann Droneaud <ydroneaud@opteya.com>
	Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit 8abaae62f3fdead8f4ce0ab46b4ab93dee39bab2)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>