md/raid10: fix conversion from RAID0 to RAID10
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [md] raid10: fix conversion from RAID0 to RAID10 (Jes Sorensen) [1150149 1173510 1194720]
Rebuild_FUZZ: 96.63%
commit-author NeilBrown <neilb@suse.de>
commit 53a6ab4d3f6d6dc87ec8f14998b4b5536ee2968c

A RAID0 array (like a LINEAR array) does not have a concept
of 'size' being the amount of each device that is in use.
Rather, as much of each device as is available is used.
So the 'size' is set to 0 and ignored.

RAID10 does have this concept and needs it to be set correctly.
So when we convert RAID0 to RAID10 we must determine the
'size' (that being the size of the first 'strip_zone' in the
RAID0), and set it correctly.

Reported-and-tested-by: Xiao Ni <xni@redhat.com>
	Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 53a6ab4d3f6d6dc87ec8f14998b4b5536ee2968c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>