md/raid5: close race between STRIPE_BIT_DELAY and batching.
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [md] raid5: close race between STRIPE_BIT_DELAY and batching (Jes Sorensen) [1150149 1173510 1194720]
Rebuild_FUZZ: 96.49%
commit-author NeilBrown <neilb@suse.de>
commit d0852df543e5aa7db34c1ad26d053782bcbf48f1

When we add a write to a stripe we need to make sure the bitmap
bit is set.  While doing that the stripe is not locked so it could
be added to a batch after which further changes to STRIPE_BIT_DELAY
and ->bm_seq are ineffective.

So we need to hold off adding to a stripe until bitmap_startwrite has
completed at least once, and we need to avoid further changes to
STRIPE_BIT_DELAY once the stripe has been added to a batch.

If a bitmap_startwrite() completes after the stripe was added to a
batch, it will not have set the bit, only incremented a counter, so no
extra delay of the stripe is needed.

	Reported-by: Shaohua Li <shli@kernel.org>
	Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit d0852df543e5aa7db34c1ad26d053782bcbf48f1)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>