md/raid5: Ensure a batch member is not handled prematurely.
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [md] raid5: Ensure a batch member is not handled prematurely (Jes Sorensen) [1150149 1173510 1194720]
Rebuild_FUZZ: 96.49%
commit-author NeilBrown <neilb@suse.de>
commit b15a9dbdbfe72848b7ed4cd3f97fe80daaf99c89

If a stripe is a member of a batch, but not the head, it must
not be handled separately from the rest of the batch.

'clear_batch_ready()' handles this requirement to some
extent but not completely.  If a member is passed to handle_stripe()
a second time it returns '0' indicating the stripe can be handled,
which is wrong.
So add an extra test.

	Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit b15a9dbdbfe72848b7ed4cd3f97fe80daaf99c89)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>