md/raid5: be more selective about distributing flags across batch.
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [md] raid5: be more selective about distributing flags across batch (Jes Sorensen) [1150149 1173510 1194720]
Rebuild_FUZZ: 96.88%
commit-author NeilBrown <neilb@suse.de>
commit 1b956f7a8f9aa63ea9644ab8c3374cf381993363

When a batch of stripes is broken up, we keep some of the flags
that were per-stripe, and copy other flags from the head to all
others.

This only happens while a stripe is being handled, so many of the
flags are irrelevant.

The "SYNC_FLAGS" (which I've renamed to make it clear there are
several) and STRIPE_DEGRADED are set per-stripe and so need to be
preserved.  STRIPE_INSYNC is the only flag that is set on the head
that needs to be propagated to all others.

For safety, add a WARN_ON if others are set, except:
 STRIPE_HANDLE - this is safe and per-stripe and we are going to set
      in several cases anyway
 STRIPE_INSYNC
 STRIPE_IO_STARTED - this is just a hint and doesn't hurt.
 STRIPE_ON_PLUG_LIST
 STRIPE_ON_RELEASE_LIST - It is a point pointless for a batched
           stripe to be on one of these lists, but it can happen
           as can be safely ignored.

	Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 1b956f7a8f9aa63ea9644ab8c3374cf381993363)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>