be2net: use offset based FW flashing for Skyhawk chip
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Vasundhara Volam <vasundhara.volam@emulex.com>
commit 70a7b5257018c518007c7212977bab6ccacd9468

While sending FW update cmds to the FW, the driver specifies the "type"
of each component that needs to be flashed. The FW then picks the offset
in the flash area at which the componnet is to be flashed. This doesn't work
when new components that the current FW doesn't recognize, need to be
flashed. Recent FWs (10.2 and above) support a scheme of FW-update wherein
the "offset" of the component in the flash area can be specified instead
of the "type". This patch uses the "offset" based FW-update mechanism and
only when it fails, it fallsback to the old "type" based update.

	Signed-off-by: Vasundhara Volam <vasundhara.volam@emulex.com>
	Signed-off-by: Sathya Perla <sathya.perla@emulex.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 70a7b5257018c518007c7212977bab6ccacd9468)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>