be2net: re-distribute SRIOV resources allowed by FW
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Vasundhara Volam <vasundhara.volam@emulex.com>
commit f285873841299e027a6159dc3d3af0d0caf578d9

When SR-IOV is enabled in the adapter, the FW distributes resources
evenly across the PF and it's VFs. This is currently done only for some
resources.

This patch adds support for a new cmd that queries the FW for the list
of resources for which the distribution is allowed and distributes them
accordingly.

	Signed-off-by: Vasundhara Volam <vasundhara.volam@emulex.com>
	Signed-off-by: Sathya Perla <sathya.perla@emulex.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit f285873841299e027a6159dc3d3af0d0caf578d9)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>