be2net: add support for spoofchk setting
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Kalesh AP <kalesh.purayil@emulex.com>
commit e7bcbd7b81a129d79f205b19f9dd96b00c7e2364

This patch adds support for spoofchk configuration for VFs.
When it is enabled, "spoof checking" is done for both MAC-address and VLAN.
For each VF, the HW ensures that the source MAC address (or vlan) of
every outgoing packet exists in the MAC-list (or vlan-list) configured
for RX filtering for that VF. If not, the packet is dropped and an error
is reported to the driver in the TX completion; this is reflected in the
"tx_spoof_check_err" ethtool counter.
This feature is supported in Skyhawk FW version 10.6.31.0 and above.

	Signed-off-by: Kalesh AP <kalesh.purayil@emulex.com>
	Signed-off-by: Sathya Perla <sathya.perla@avagotech.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit e7bcbd7b81a129d79f205b19f9dd96b00c7e2364)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>